파일분리 : AnalysisSolveProblem 분리
